+++ src/main/java/com/plus/domain/review/repository/ReviewRepository.java
@@ -1,8 +1,8 @@
-package com.plus.domain.draw.repository;
+package com.plus.domain.review.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.plus.domain.draw.entity.Review;
+import com.plus.domain.review.entity.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 }
